requirements.txt
# FastAPI & Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# AI & LLM
mistralai==0.0.12
langchain-text-splitters==0.0.1

# Document Processing
python-docx==0.8.11
python-pptx==0.6.21
reportlab==4.0.7
PyPDF2==3.0.1

# Text Processing & Utilities
python-bidi==0.4.2
arabic-reshaper==0.5.2

# System & Helpers
requests==2.31.0
aiofiles==23.2.1
